+++ docs/.vuepress/config.js
@@ -24,10 +24,6 @@ module.exports = {
         link: '/api/'
       },
       {
-        text: "License",
-        link: '/license.html'
-      },
-      {
         text: "Log Out",
         link: '/logout.html'
       }
@@ -109,7 +105,6 @@ module.exports = {
           ]
         }
       ],
-      '/license': [],
       '/guide/': [
         '/guide/',
         {