+++ src/prefect/utilities/context.py
@@ -1,5 +1,3 @@
-# Licensed under LICENSE.md; also available at https://www.prefect.io/licenses/beta-eula
-
 """
 This module implements the Prefect context that is available when tasks run.